tools/prompt_injection_tester/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-injection-tester"
version = "2.0.0"
description = "Modern CLI for automated prompt injection testing of LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "CC BY-SA 4.0"}
authors = [
    {name = "AI LLM Red Team Handbook"}
]
keywords = [
    "llm",
    "security",
    "prompt-injection",
    "red-team",
    "testing",
    "cli",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.9.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
prompt-injection-tester = "prompt_injection_tester.cli:main"
pit = "pit.app:cli_main"

[project.urls]
"Homepage" = "https://github.com/example/ai-llm-red-team-handbook"
"Documentation" = "https://github.com/example/ai-llm-red-team-handbook/tree/main/tools/prompt_injection_tester"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
